Adjust Packaging
Issue gh-9289
