+++ config/src/main/java/org/springframework/security/config/annotation/method/configuration/MethodSecurityConfiguration.java
@@ -45,19 +45,19 @@ import org.springframework.context.annotation.Role;
 import org.springframework.core.annotation.AnnotatedElementUtils;
 import org.springframework.core.annotation.AnnotationAttributes;
 import org.springframework.core.type.AnnotationMetadata;
-import org.springframework.security.access.annotation.Jsr250AuthorizationManager;
+import org.springframework.security.authorization.method.Jsr250AuthorizationManager;
 import org.springframework.security.access.annotation.Secured;
-import org.springframework.security.access.annotation.SecuredAuthorizationManager;
+import org.springframework.security.authorization.method.SecuredAuthorizationManager;
 import org.springframework.security.access.expression.method.DefaultMethodSecurityExpressionHandler;
 import org.springframework.security.access.expression.method.MethodSecurityExpressionHandler;
-import org.springframework.security.access.intercept.aopalliance.AuthorizationMethodInterceptor;
-import org.springframework.security.access.method.AuthorizationManagerMethodAfterAdvice;
-import org.springframework.security.access.method.AuthorizationManagerMethodBeforeAdvice;
-import org.springframework.security.access.method.AuthorizationMethodAfterAdvice;
-import org.springframework.security.access.method.AuthorizationMethodBeforeAdvice;
-import org.springframework.security.access.method.DelegatingAuthorizationMethodAfterAdvice;
-import org.springframework.security.access.method.DelegatingAuthorizationMethodBeforeAdvice;
-import org.springframework.security.access.method.MethodAuthorizationContext;
+import org.springframework.security.authorization.method.AuthorizationMethodInterceptor;
+import org.springframework.security.authorization.method.AuthorizationManagerMethodAfterAdvice;
+import org.springframework.security.authorization.method.AuthorizationManagerMethodBeforeAdvice;
+import org.springframework.security.authorization.method.AuthorizationMethodAfterAdvice;
+import org.springframework.security.authorization.method.AuthorizationMethodBeforeAdvice;
+import org.springframework.security.authorization.method.DelegatingAuthorizationMethodAfterAdvice;
+import org.springframework.security.authorization.method.DelegatingAuthorizationMethodBeforeAdvice;
+import org.springframework.security.authorization.method.MethodAuthorizationContext;
 import org.springframework.security.access.prepost.PostAuthorize;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.authorization.method.PostAuthorizeAuthorizationManager;
