+++ config/src/test/java/org/springframework/security/config/annotation/method/configuration/MethodSecurityConfigurationTests.java
@@ -35,10 +35,10 @@ import org.springframework.security.access.annotation.BusinessService;
 import org.springframework.security.access.annotation.ExpressionProtectedBusinessServiceImpl;
 import org.springframework.security.access.expression.method.DefaultMethodSecurityExpressionHandler;
 import org.springframework.security.access.expression.method.MethodSecurityExpressionHandler;
-import org.springframework.security.access.method.AuthorizationManagerMethodBeforeAdvice;
-import org.springframework.security.access.method.AuthorizationMethodAfterAdvice;
-import org.springframework.security.access.method.AuthorizationMethodBeforeAdvice;
-import org.springframework.security.access.method.MethodAuthorizationContext;
+import org.springframework.security.authorization.method.AuthorizationManagerMethodBeforeAdvice;
+import org.springframework.security.authorization.method.AuthorizationMethodAfterAdvice;
+import org.springframework.security.authorization.method.AuthorizationMethodBeforeAdvice;
+import org.springframework.security.authorization.method.MethodAuthorizationContext;
 import org.springframework.security.authorization.AuthorizationDecision;
 import org.springframework.security.authorization.AuthorizationManager;
 import org.springframework.security.config.test.SpringTestRule;
