+++ core/src/main/java/org/springframework/security/authorization/method/AuthorizationMethodBeforeAdvice.java
@@ -14,26 +14,24 @@
  * limitations under the License.
  */
 
-package org.springframework.security.access.method;
+package org.springframework.security.authorization.method;
 
 import java.util.function.Supplier;
-
-import org.aopalliance.intercept.MethodInvocation;
 
 import org.springframework.aop.ClassFilter;
 import org.springframework.aop.Pointcut;
 import org.springframework.security.core.Authentication;
 
 /**
- * An Authorization advice that can determine if an {@link Authentication} has access to
- * the returned object from the {@link MethodInvocation}. The {@link #getMethodMatcher()}
- * describes when the advice applies for the method.
+ * An advice which can determine if an {@link Authentication} has access to the {@link T}
+ * object. The {@link #getMethodMatcher()} describes when the advice applies for the
+ * method.
  *
  * @param <T> the type of object that the authorization check is being done one.
  * @author Evgeniy Cheban
  * @since 5.5
  */
-public interface AuthorizationMethodAfterAdvice<T> extends Pointcut {
+public interface AuthorizationMethodBeforeAdvice<T> extends Pointcut {
 
 	/**
 	 * Returns the default {@link ClassFilter}.
@@ -45,17 +43,10 @@ public interface AuthorizationMethodAfterAdvice<T> extends Pointcut {
 	}
 
 	/**
-	 * Determines if an {@link Authentication} has access to the returned object from the
-	 * {@link MethodInvocation}.
+	 * Determines if an {@link Authentication} has access to the {@link T} object.
 	 * @param authentication the {@link Supplier} of the {@link Authentication} to check
 	 * @param object the {@link T} object to check
-	 * @param returnedObject the returned object from the {@link MethodInvocation} to
-	 * check
-	 * @return the <code>Object</code> that will ultimately be returned to the caller (if
-	 * an implementation does not wish to modify the object to be returned to the caller,
-	 * the implementation should simply return the same object it was passed by the
-	 * <code>returnedObject</code> method argument)
 	 */
-	Object after(Supplier<Authentication> authentication, T object, Object returnedObject);
+	void before(Supplier<Authentication> authentication, T object);
 
 }
