+++ core/src/main/java/org/springframework/security/authorization/method/MethodAuthorizationContext.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.security.access.method;
+package org.springframework.security.authorization.method;
 
 import org.aopalliance.intercept.MethodInvocation;
