+++ core/src/test/java/org/springframework/security/authorization/method/SecuredAuthorizationManagerTests.java
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package org.springframework.security.access.annotation;
+package org.springframework.security.authorization.method;
 
 import java.util.function.Supplier;
 
 import org.junit.Test;
 
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.access.intercept.method.MockMethodInvocation;
-import org.springframework.security.access.method.MethodAuthorizationContext;
 import org.springframework.security.authentication.TestAuthentication;
 import org.springframework.security.authentication.TestingAuthenticationToken;
 import org.springframework.security.authorization.AuthorizationDecision;
